+++ src/main/java/com/boardgo/domain/boardgame/service/BoardGameQueryServiceV1.java
@@ -3,8 +3,8 @@ package com.boardgo.domain.boardgame.service;
 import com.boardgo.domain.boardgame.controller.request.BoardGameSearchRequest;
 import com.boardgo.domain.boardgame.repository.BoardGameRepository;
 import com.boardgo.domain.boardgame.repository.projection.BoardGameSearchProjection;
-import com.boardgo.domain.boardgame.repository.response.BoardGameSearchResponse;
-import com.boardgo.domain.boardgame.repository.response.GenreSearchResponse;
+import com.boardgo.domain.boardgame.service.response.BoardGameSearchResponse;
+import com.boardgo.domain.boardgame.service.response.GenreSearchResponse;
 import com.boardgo.domain.mapper.BoardGameMapper;
 import java.util.List;
 import java.util.Map;
