+++ src/main/java/com/boardgo/domain/boardgame/service/BoardGameQueryUseCase.java
@@ -1,7 +1,7 @@
 package com.boardgo.domain.boardgame.service;
 
 import com.boardgo.domain.boardgame.controller.request.BoardGameSearchRequest;
-import com.boardgo.domain.boardgame.repository.response.BoardGameSearchResponse;
+import com.boardgo.domain.boardgame.service.response.BoardGameSearchResponse;
 import org.springframework.data.domain.Page;
 
 public interface BoardGameQueryUseCase {
