+++ src/main/java/com/boardgo/domain/boardgame/service/response/BoardGameListResponse.java
@@ -1,3 +1,3 @@
-package com.boardgo.domain.boardgame.repository.response;
+package com.boardgo.domain.boardgame.service.response;
 
 public record BoardGameListResponse(Long boardGameId, String title, String thumbnail) {}
