+++ src/main/java/com/boardgo/domain/boardgame/service/response/GenreSearchResponse.java
@@ -1,3 +1,3 @@
-package com.boardgo.domain.boardgame.repository.response;
+package com.boardgo.domain.boardgame.service.response;
 
 public record GenreSearchResponse(Long id, String genre) {}
