+++ src/main/java/com/boardgo/domain/mapper/BoardGameGenreMapper.java
@@ -1,7 +1,7 @@
 package com.boardgo.domain.mapper;
 
 import com.boardgo.domain.boardgame.repository.projection.GenreSearchProjection;
-import com.boardgo.domain.boardgame.repository.response.GenreSearchResponse;
+import com.boardgo.domain.boardgame.service.response.GenreSearchResponse;
 import org.mapstruct.Mapper;
 import org.mapstruct.factory.Mappers;
