+++ src/main/java/com/boardgo/domain/mapper/BoardGameMapper.java
@@ -4,10 +4,10 @@ import com.boardgo.domain.boardgame.controller.request.BoardGameCreateRequest;
 import com.boardgo.domain.boardgame.entity.BoardGameEntity;
 import com.boardgo.domain.boardgame.repository.projection.BoardGameByMeetingIdProjection;
 import com.boardgo.domain.boardgame.repository.projection.BoardGameSearchProjection;
-import com.boardgo.domain.boardgame.repository.response.BoardGameByMeetingIdResponse;
-import com.boardgo.domain.boardgame.repository.response.BoardGameListResponse;
-import com.boardgo.domain.boardgame.repository.response.BoardGameSearchResponse;
-import com.boardgo.domain.boardgame.repository.response.GenreSearchResponse;
+import com.boardgo.domain.boardgame.service.response.BoardGameListResponse;
+import com.boardgo.domain.boardgame.service.response.BoardGameSearchResponse;
+import com.boardgo.domain.boardgame.service.response.GenreSearchResponse;
+import com.boardgo.domain.meeting.service.response.BoardGameByMeetingIdResponse;
 import java.util.List;
 import java.util.Map;
 import java.util.Set;
