+++ src/main/java/com/boardgo/domain/meeting/service/MeetingQueryServiceV1.java
@@ -3,18 +3,18 @@ package com.boardgo.domain.meeting.service;
 import com.boardgo.common.exception.CustomNoSuchElementException;
 import com.boardgo.common.utils.SecurityUtils;
 import com.boardgo.domain.boardgame.repository.BoardGameRepository;
-import com.boardgo.domain.boardgame.repository.response.BoardGameByMeetingIdResponse;
 import com.boardgo.domain.mapper.MeetingMapper;
 import com.boardgo.domain.meeting.controller.request.MeetingSearchRequest;
 import com.boardgo.domain.meeting.repository.MeetingLikeRepository;
 import com.boardgo.domain.meeting.repository.MeetingRepository;
 import com.boardgo.domain.meeting.repository.projection.MeetingDetailProjection;
 import com.boardgo.domain.meeting.repository.projection.MeetingSearchProjection;
-import com.boardgo.domain.meeting.repository.response.MeetingDetailResponse;
-import com.boardgo.domain.meeting.repository.response.MeetingSearchResponse;
+import com.boardgo.domain.meeting.service.response.BoardGameByMeetingIdResponse;
+import com.boardgo.domain.meeting.service.response.MeetingDetailResponse;
+import com.boardgo.domain.meeting.service.response.MeetingSearchResponse;
+import com.boardgo.domain.meeting.service.response.UserParticipantResponse;
 import com.boardgo.domain.review.repository.ReviewRepository;
 import com.boardgo.domain.user.repository.UserRepository;
-import com.boardgo.domain.user.repository.response.UserParticipantResponse;
 import java.util.List;
 import java.util.Map;
 import java.util.Objects;
