+++ src/main/java/com/boardgo/domain/meeting/service/MeetingQueryUseCase.java
@@ -1,8 +1,8 @@
 package com.boardgo.domain.meeting.service;
 
 import com.boardgo.domain.meeting.controller.request.MeetingSearchRequest;
-import com.boardgo.domain.meeting.repository.response.MeetingDetailResponse;
-import com.boardgo.domain.meeting.repository.response.MeetingSearchResponse;
+import com.boardgo.domain.meeting.service.response.MeetingDetailResponse;
+import com.boardgo.domain.meeting.service.response.MeetingSearchResponse;
 import org.springframework.data.domain.Page;
 
 public interface MeetingQueryUseCase {
