+++ src/main/java/com/boardgo/domain/meeting/service/response/UserParticipantResponse.java
@@ -1,4 +1,4 @@
-package com.boardgo.domain.user.repository.response;
+package com.boardgo.domain.meeting.service.response;
 
 import com.boardgo.domain.meeting.entity.enums.ParticipantType;
