+++ src/main/java/com/boardgo/domain/user/repository/UserDslRepository.java
@@ -1,7 +1,7 @@
 package com.boardgo.domain.user.repository;
 
+import com.boardgo.domain.meeting.service.response.UserParticipantResponse;
 import com.boardgo.domain.user.repository.projection.PersonalInfoProjection;
-import com.boardgo.domain.user.repository.response.UserParticipantResponse;
 import java.util.List;
 
 public interface UserDslRepository {
